add base installation of hephy, and metallb
This is a basic HelmRelease that installs a tool called Hephy Workflow

It will need a Load Balancer, so in the Home Lab we can use MetalLB
(In the cloud, Load Balancers are plentiful and you can work without it)

Hephy Workflow is like a Dokku/Heroku instance for your Kubernetes
You can push git to deploy to it, and it has backups, and a database!
And login, and `deis ps` and users, and Experimental Native Ingress.

(We will make Experimental Native Ingress work with Traefik down on the
line! Watch this space and stay tuned for more :look:!)