add metallb as an "app profile"
This really messes with the nice directory structure that Jakub
introduced. I wonder if we should have a separate directory for app
profiles or app instances? This metallb app does not go on production or
staging, and it will only be used on one cluster. (The same is true for
Hephy Workflow next.)