Hephy Workflow blog and Workflow instance
My database and backup of this Hephy Workflow instance will launch a
blog, that is the Hephy Workflow blog. (Yours may do something else.)

I have created an IngressRoute for http://blog.teamhephy.info here
and arranged for it to be TLS terminated another way. This is a lazy
integration between Hephy Workflow and the Traefik ingress controller.

(We can try to show a better one later on -- the Experimental Native
Ingress support is compatible with the Traefik Kubernetes Ingress mode!)