"cluster-api" aks and gke clusters
Introducing, alternatively / alternating as the Staging and Production
instances, we have these two clusters which can be an Azure AKS cluster
or a GKE cluster, from Google. I don't think it matters which we use.

We will bootstrap Flux onto the moo-cluster then onto both Staging and
Production clusters from here, setting spec.kubeConfig.secretRef.name to
the name of an encrypted secret where we'll (temporarily or not) store a
copy of the kubeConfig for each cluster that we have to manage.