cluster-api secrets
Including a script for generating a kubeconfig on GKE, which is a bit
more involved than just calling a CLI "get-credentials" method, like I
believe that I found I actually could do in Azure pretty handily.

These are SOPS-encrypted with a GPG cluster that I provisioned on my
local network, the moo-cluster runs its leader as a machine called "moo"
on the network "turkey.local" so it's cluster-api.turkey.local :+1:

I have fetched two valid kubeconfig files here and anyone can refresh
them if they are ever invalidated. The public key is in the file here
also, and a SOPS config .sops.yaml that tells the sops CLI how to
encrypt files, and what key to use.

So secrets can be written by anyone with no special access, as long as
they add the public key, and if they can get their pull request merged.
The SOPS secret storage is transparent about the structure of a secret.
(You could contribute a pull request to get your own cluster added here,
it would be safe enough!)