no longer depend on the default ingress class
it looks like "internal" is being assigned as the default ingress class
due to some change. This is always how I expected it to work, but I
don't think it ever actually worked before... anyway, we can no longer
pick up ingresses that are missing an ingressclass with Traefik if the
default class is assigned instead. So let's declare an ingress class.