set us up for two more cert renewals
vpn.hephy.pro looks like it's actually set up wrong! (it is a public
service so its traefik config is part of itself, but it is not set up
properly for tls renewal)

It should always point at the public ingress class. The cert issuers are
all configured to use that ingress class. So we can make ingresses for
cert http-01 challenges public without making the service public, and
this does not require anyone to flip a switch in order to activate 👍✨
the certificates will just stay fresh, and it will be fine as long as
there is no flux sync configured to overwrite them with an older version
whenever they renewed.