domains scheme
In the cozystack docs, a domain name is partitioned into tenants, which
then has clusters. A cluster is:

cluster-1.test.moomboo.space

The `test.moomboo.space` is the name of the tenant, and `cluster-1`
lives in the tenant. Another cluster in the root tenant, if the root
tenant is `howard.moomboo.space` might be

`prod.howard.moomboo.space`

where the cluster name is `prod` and the tenant is `moomboo.space`

These are all hosted on a cluster named `moomboo.space` or in this case,
`cozy.local` - these do not have to be continuous, and of course any
tenant or cluster name can belong to any domain name, if you decide it.

The Kamaji stuff is described as vaguely hierarchical, with everything
in a tenant visible to the parent tenant.

Anyway, there should not be a cluster named `test.moomboo.space` because
it is the name of a tenant. And tenants may have more than one cluster!

This traefik route will enable us to create clusters with arbitrary
names, and they will get routed in from the outside.

(If you want to use a subdomain for ingress, it will have to be
something else, because these subdomains are all going to go to the root
ingress controller for CAPI clusters.)

The goal for today is to migrate `howard.moomboo.space` or `moo` - or
ideally both!

We have services like harbor, ingress, traefik routers, dex, and other
services that depend on them in these clusters.

(There are also a lot of services that we could delete, we don't need
all of them. We've done a lot of demos over the years, some are still
running! Almost all still have leftover ingress rules that we should
probably clean up, or memorialize somehow.)

We need many of those services to stay online during the migration, or
at least for any interruption to be brief and limited, with certainty
about when it will be able to be restored. Any necessary interruption
should be minimized because these services are depended upon to reach
other services, and that makes them "business-critical" even if business
people don't think of them at all!

I use Dex to authenticate to all of the clusters, in the old model.

Any vcluster can easily be set up to work with Dex. So far the Cozystack
model hasn't added support for any Dex integration with the Kamaji
control planes, but it looks like it would be trivial to add it.

Anyway, we haven't decided where we're gonna put the vclusters, but I
think for now it's going to all go on howard.moomboo.space

The old cluster `howard.moomboo.space` is replaced with a KubeVirt
cluster that runs all the same workloads as the old vcluster

It runs CAPI internally and it runs the CAPI provider vcluster, which
has put out a new release that we should try.

It runs only one cluster, `limnocentral` which will be renamed to

`limnocentral.howard.moomboo.space`

This cluster breaks the rule because there should be no other Kube
clusters in the root tenant. The root tenant is named
howard.moomboo.space and the one kube cluster in this tenant is called
howard. All other kube clusters will be a 4th-level domain name or on a
new domain. This simplifies host based routing for us dramatically.

We also have `kingdonb.dev` and `yaateeh.dev` that were renewed.

We can add more domains later, but we have to get these ones migrated so
we can borrow the nodes they were running on, when the last of critical
services have been migrated! First I think I should try to migrate Dex.

Signed-off-by: Kingdon Barrett <kingdon+github@tuesdaystudios.com>